+++ examples/app/src/main.rs
@@ -3,11 +3,11 @@ use trillium_frontend::frontend;
 use trillium_logger::Logger;
 use trillium_smol::ClientConfig;
 
-pub fn main() {
+fn main() {
     env_logger::init();
     trillium_smol::run((
         Logger::new(),
-        frontend!("./examples/client")
+        frontend!("./client")
             .with_client(Client::new(ClientConfig::default()))
             .with_index_file("index.html"),
     ));